server/celery_workers/requirements.txt
celery[redis]
python-decouple==3.5
pymongo
dnspython==2.1.0